requirements.txt
google.generativeai==0.4.1
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.32.1
streamlit-option-menu==0.3.12
streamlit-authenticator==0.3.1
